scripts/roboflow_rfdetr/requirements.txt
# RF-DETR Vineyard Detection Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# RF-DETR model
rfdetr>=1.4.0

# Data handling
pycocotools>=2.0.6
pyyaml>=6.0

# Visualization and utilities
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional but recommended
# CUDA support (if using GPU)
# Install appropriate version from https://pytorch.org/get-started/locally/
